added auth (#4)
* initial commit

* chore: added logs to service

* refactor: server

* chore: changed handler responses for errors

* ci: added registry publish workflow

* nothing much

* chore: added grafana agent

* chore: added config template

* chore: shifter http server to native lib

* backup

* chore: added more middlewares

* fix: middleware order

* chore: removed `start_time` from log records

* chore: migrated server handler to `main` pkg

* backup: added separate router pkg

* chore: improved http router

* chore: connected the 3 layers

* chore: segregated router and handlers

* chore: removed old handler pkg

* chore: renamed body to options

* refactor: renamed `todo` svc to `record` svc

* cleanup: removed unrequired pkgs

* chore: added `list` handler

* chore: added `get` handler for `records` svc

* feat: completed all handlers

* chore: added unit tests for `db` layer

* Merge branch 'main' into dev

* refactor: removed server pkg

* chore: refreshed `go.sum`

* Merge branch 'main' into dev

* refactor: db tests

* feat: shifted to connection pooling

* chore: (hopefully) fixed connection pooling

* chore: refactor `docker-compose.yml`

* chore: added db migrations wkflw

* chore: renamed docker-compose for local use

* chore: added service layer unit tests

* chore: updated READMEs

* chore: updated readme

* chore: updated readme

* chore: updated readme

* chore: updated readme

* Merge branch 'main' into dev

* chore: removed unrequired logger from db instance

* chore: removed unrequired logger from db layer

* chore: shifted entire svc to use same gorm conn

* chore: renamed `deploy` to `setup` pkg

* refactor: renamed `handlers` pkg to `handler`

* chore: shifted `handler` pkg to common `Handler` interface

* migration: updated model

+++ service/service.go
@@ -1,3 +1,4 @@
+//go:generate mockgen -destination=mock.go -source=service.go -package=service
 package service
 
 import (
@@ -6,13 +7,14 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/mrinalwahal/service/db"
+	"github.com/mrinalwahal/service/model"
 )
 
 type Service interface {
-	Create(context.Context, *CreateOptions) (*db.Record, error)
-	List(context.Context, *ListOptions) ([]*db.Record, error)
-	Get(context.Context, uuid.UUID) (*db.Record, error)
-	Update(context.Context, uuid.UUID, *UpdateOptions) (*db.Record, error)
+	Create(context.Context, *CreateOptions) (*model.Record, error)
+	List(context.Context, *ListOptions) ([]*model.Record, error)
+	Get(context.Context, uuid.UUID) (*model.Record, error)
+	Update(context.Context, uuid.UUID, *UpdateOptions) (*model.Record, error)
 	Delete(context.Context, uuid.UUID) error
 }
 
@@ -50,19 +52,26 @@ type service struct {
 	logger *slog.Logger
 }
 
-func (s *service) Create(ctx context.Context, options *CreateOptions) (*db.Record, error) {
+func (s *service) Create(ctx context.Context, options *CreateOptions) (*model.Record, error) {
 	s.logger.LogAttrs(ctx, slog.LevelDebug, "creating a new record",
 		slog.String("function", "create"),
 	)
+	if options == nil {
+		return nil, ErrInvalidArguments
+	}
 	return s.db.Create(ctx, &db.CreateOptions{
-		Title: options.Title,
+		Title:  options.Title,
+		UserID: options.UserID,
 	})
 }
 
-func (s *service) List(ctx context.Context, options *ListOptions) ([]*db.Record, error) {
+func (s *service) List(ctx context.Context, options *ListOptions) ([]*model.Record, error) {
 	s.logger.LogAttrs(ctx, slog.LevelDebug, "listing all records",
 		slog.String("function", "list"),
 	)
+	if options == nil {
+		return nil, ErrInvalidArguments
+	}
 	return s.db.List(ctx, &db.ListOptions{
 		Title:          options.Title,
 		Skip:           options.Skip,
@@ -72,17 +81,26 @@ func (s *service) List(ctx context.Context, options *ListOptions) ([]*db.Record,
 	})
 }
 
-func (s *service) Get(ctx context.Context, ID uuid.UUID) (*db.Record, error) {
+func (s *service) Get(ctx context.Context, ID uuid.UUID) (*model.Record, error) {
 	s.logger.LogAttrs(ctx, slog.LevelDebug, "retrieving a record",
 		slog.String("function", "get"),
 	)
+	if ID == uuid.Nil {
+		return nil, ErrInvalidArguments
+	}
 	return s.db.Get(ctx, ID)
 }
 
-func (s *service) Update(ctx context.Context, ID uuid.UUID, options *UpdateOptions) (*db.Record, error) {
+func (s *service) Update(ctx context.Context, ID uuid.UUID, options *UpdateOptions) (*model.Record, error) {
 	s.logger.LogAttrs(ctx, slog.LevelDebug, "updating a record",
 		slog.String("function", "update"),
 	)
+	if ID == uuid.Nil {
+		return nil, ErrInvalidArguments
+	}
+	if options == nil {
+		return nil, ErrInvalidArguments
+	}
 	return s.db.Update(ctx, ID, &db.UpdateOptions{
 		Title: options.Title,
 	})
@@ -92,5 +110,8 @@ func (s *service) Delete(ctx context.Context, ID uuid.UUID) error {
 	s.logger.LogAttrs(ctx, slog.LevelDebug, "deleting a record",
 		slog.String("function", "delete"),
 	)
+	if ID == uuid.Nil {
+		return ErrInvalidArguments
+	}
 	return s.db.Delete(ctx, ID)
 }
